fix vesc fault code data type

+++ vesc_driver/include/vesc_driver/vesc_interface.h
@@ -85,7 +85,7 @@ namespace vesc_driver {
         uint32_t tacho;
         uint32_t tacho_absolute;
         bool direction;
-        int32_t fault_code;
+        uint8_t fault_code;
     };
 
 /**
